refactor: rename Alerts into AlertsLegacy
