+++ packages/core/test/Core.spec.js
@@ -15,7 +15,7 @@ const MocksMock = require("./mocks/Mocks.mock.js");
 const ServerMocks = require("./server/Server.mocks.js");
 const PluginsMocks = require("./plugins/Plugins.mocks.js");
 const ConfigMocks = require("./Config.mocks.js");
-const AlertsMocks = require("./Alerts.mocks.js");
+const AlertsMocks = require("./AlertsLegacy.mocks.js");
 const LoadersMocks = require("./Loaders.mocks.js");
 const FilesLoaderMocks = require("./files-loader/FilesLoader.mocks.js");
 const ScaffoldMocks = require("./scaffold/Scaffold.mocks.js");
